rename the file after testing multiple pictures
